// apps/enterprise-service-management/src/generated/models/workspace-model.ts
import type { Person } from './person-model';

export const WorkspaceActiveToLabel = {
  'Active': 'Active',
  'Inactive': 'Inactive',
  'Suspended': 'Suspended'
} as const;
export type WorkspaceActive = keyof typeof WorkspaceActiveToLabel;

export interface Workspace {
  /**
   * @displayName Workspace ID
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName WorkspaceName
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  workspaceName: string;
  /**
   * @displayName Status
   * @validationRule Required for create/update operations
   */
  active: WorkspaceActive;
  /**
   * @displayName AgentWorkspaceEnabled
   * @validationRule Required for create/update operations
   */
  agentWorkspaceEnabled: boolean;
  /**
   * @displayName CreatedAt
   * @validationRule Required for create/update operations
   */
  createdAt: string;
  /**
   * @displayName CreatedBy
   */
  createdBy?: Pick<Person, 'id' | 'displayName'>;
  /**
   * @displayName DefaultCalendarId
   */
  defaultCalendarId?: string;
  /**
   * @displayName DefaultSiteId
   */
  defaultSiteId?: string;
  /**
   * @displayName DefaultTimeZone
   * @validationRule Required for create/update operations
   */
  defaultTimeZone: string;
  /**
   * @displayName Description
   * @validationRule Required for create/update operations
   */
  description: string;
  /**
   * @displayName Icon
   * @validationRule Required for create/update operations
   */
  icon: string;
  /**
   * @displayName OwnerPersonId
   */
  ownerPersonId?: Pick<Person, 'id' | 'displayName'>;
  /**
   * @displayName PrimaryColor
   * @validationRule Required for create/update operations
   */
  primaryColor: string;
  /**
   * @displayName RequesterPortalEnabled
   * @validationRule Required for create/update operations
   */
  requesterPortalEnabled: boolean;
  /**
   * @displayName RequestNumberFormat
   * @validationRule Required for create/update operations
   */
  requestNumberFormat: string;
  /**
   * @displayName RequestNumberPrefix
   * @validationRule Required for create/update operations
   */
  requestNumberPrefix: string;
  /**
   * @displayName UpdatedAt
   * @validationRule Required for create/update operations
   */
  updatedAt: string;
  /**
   * @displayName UpdatedBy
   */
  updatedBy?: Pick<Person, 'id' | 'displayName'>;
  /**
   * @displayName WorkspaceCode
   * @validationRule Required for create/update operations
   */
  workspaceKey: string;
}

export const _Workspace = 'Workspace' as const;